The show mirror_session CLI command is not working on es9618xx.
- Why I did it:
Code defect for unknown ASIC.
The queue used by the mirror session is displayed according to the type of ASIC.
It does not take into account that when encountering a new ASIC, the CLI will
crash because the queue variable is not defined.

- What I did:
If an unsupported ASIC is encountered, "show mirror_session" will maintain the
same format as the community without showing the queue.

- How I verified it:
The "show mirror_session" is working on the es9618xx device.