QoS template files were not being applied on the Xsight ES9618xx device.
What I did:
The original implementation applied QoS template files only when running
with ecSAI. The condition was updated to apply the QoS template files by
default, except when running with the community Broadcom SAI. As a
result, the Xsight ES9618xx device is now able to load the QoS templates
as well.

Why I did it:
The lack of QoS template files on the Xsight ES9618xx caused failures in
the pfc_config and ecn_config_update test cases.